requirements.txt
streamlit
google-generativeai
python-dotenv
graphviz
streamlit-quill
PyMuPDF
python-docx
Pillow
reportlab
